Worked on the controllers, services, DAO for playlist & Video

+++ server/src/main/java/com/transteven/multiplaylist/video/VideoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VideoRepository extends JpaRepository<Video, Long>{}
+public interface VideoRepository extends JpaRepository<Video, Integer> {}
